- Silence a bunch of compile warnings on Mac OS X. - Fix bugid: 2013-02-07-001
bk: 511b9689QXdF-VKHIsd5b1SqEE4I6g

+++ src/bk.c
@@ -1342,8 +1342,8 @@ write_log(char *file, char *format, ...)
 	gettimeofday(&tv, 0);
 	sprintf(nformat, "%c%s %lu.%06lu %s %s: %*s%s\n",
 	    log_versions[LOGVER],
-	    sccs_user(), tv.tv_sec, tv.tv_usec, milli(), bk_vers,
-	    indent(), "", format);
+	    sccs_user(), tv.tv_sec, (unsigned long)tv.tv_usec,
+	    milli(), bk_vers, indent(), "", format);
 	va_start(ap, format);
 	vfprintf(f, nformat, ap);
 	va_end(ap);
